+++ .gitignore
@@ -30,3 +30,4 @@ hs_err_pid*
 *.class
 *.log
 *.properties
+*.springBeans
