+++ pom.xml
@@ -3,7 +3,7 @@
     xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>org.zoophy</groupId>
+    <groupId>edu.asu.zoophy</groupId>
     <artifactId>zoophy-rest-service</artifactId>
     <version>0.0.2</version>
