--- src/main/java/com/zoophy/security/Parameter.java
@@ -1,12 +1,0 @@
-package com.zoophy.security;
-
-/**
- * Types of allowed String parameters for services
- * @author developerdemetri
- */
-public enum Parameter {
-	ACCESSION,
-	LUCENE_QUERY, 
-	EMAIL,
-	JOB_NAME
-}
