+++ src/main/java/edu/asu/zoophy/pipeline/ZooPhyRunner.java
@@ -1,0 +1,60 @@
+package edu.asu.zoophy.pipeline;
+
+import java.util.Collections;
+import java.util.List;
+import java.util.Set;
+import java.util.concurrent.ConcurrentHashMap;
+
+/**
+ * Responsible for running ZooPhy jobs
+ * @author devdemetri
+ */
+public class ZooPhyRunner {
+	
+	private ZooPhyJob job;
+	private ZooPhyMailer mailer;
+	
+	private static Set<String> ids = Collections.newSetFromMap(new ConcurrentHashMap<String, Boolean>());
+	
+	public ZooPhyRunner(String replyEmail) {
+		job = new ZooPhyJob(generateID(),null,replyEmail);
+	}
+
+	public ZooPhyRunner(String replyEmail, String jobName) {
+		job = new ZooPhyJob(generateID(),jobName,replyEmail);
+	}
+	
+	/**
+	 * Runs the ZooPhy pipeline on the given Accessions
+	 * @param accessions
+	 * @throws PipelineException
+	 */
+	public void runZooPhy(List<String> accessions) throws PipelineException {
+		try {
+			mailer = new ZooPhyMailer(job);
+			mailer.sendStartEmail();
+			//TODO: add rest of pipeline
+			// mailer.sendSuccessEmail();
+		}
+		catch (PipelineException pe) {
+			mailer.sendFailureEmail(pe.getUserMessage());
+		}
+		catch (Exception e) {
+			mailer.sendFailureEmail("Server Error");
+		}
+	}
+	
+	/**
+	 * Generates a UUID to be used as a jobID
+	 * @return Unused UUID
+	 */
+	private static String generateID() {
+		String id  = java.util.UUID.randomUUID().toString();
+		while (ids.contains(id)) {
+			id  = java.util.UUID.randomUUID().toString();
+		}
+		ids.add(id);
+		return id;
+	}
+
+}
