+++ src/main/java/edu/asu/zoophy/security/Parameter.java
@@ -1,0 +1,12 @@
+package edu.asu.zoophy.security;
+
+/**
+ * Types of allowed String parameters for services
+ * @author developerdemetri
+ */
+public enum Parameter {
+	ACCESSION,
+	LUCENE_QUERY, 
+	EMAIL,
+	JOB_NAME
+}
